feat: replace certificate-store with abstraction
