+++ agent-control/src/opamp/remote_config/validators/signature.rs
@@ -1,6 +1,5 @@
 mod certificate;
 mod certificate_fetcher;
-mod certificate_store;
 pub mod validator;
 mod verifier;
 pub use certificate::public_key_fingerprint;
